embed_assets.py now imports images too
Still need to call it from editor in pre build step and generate a c array from the image data

+++ DuskEngine/src/Platform/OpenGL/Texture.h
@@ -43,7 +43,7 @@ namespace DuskEngine
 		uint32_t Width;
 		uint32_t Height;
 		uint32_t Channels;
-		size_t DataSize;
+		uint64_t DataSize;
 	};
 
 	struct Texture : Asset, OpenGLResource
